Implement generic Boards and Tiles
Introduce generic class Tile. Boards are also now generic and containt a 2d arraylist of Tile<T>.
Each unique implementation of the game needs to pass Board a combine lambda function that defines how to combine two Tiles.
This is meant to allow for more modularity. Instead of just 2048 with ints and a combination function that just adds two ints, one can create a game of say Strings, and define a combination function as concatenating the strings, or as the least common substring of the two strings.
A big reason for doing this was to just practice Java generics.
Currently, slide logic is not perfect, but the gridview is updating!